+++ src/main/java/com/weeklyMission/client/VoucherType.java
@@ -1,9 +1,9 @@
 package com.weeklyMission.client;
 
 import com.weeklyMission.console.ConsoleIO;
-import com.weeklyMission.Voucher.domain.FixedAmountVoucher;
-import com.weeklyMission.Voucher.domain.PercentDiscountVoucher;
-import com.weeklyMission.Voucher.domain.Voucher;
+import com.weeklyMission.voucher.domain.FixedAmountVoucher;
+import com.weeklyMission.voucher.domain.PercentDiscountVoucher;
+import com.weeklyMission.voucher.domain.Voucher;
 import com.weeklyMission.exception.IncorrectInputException;
 import java.util.Arrays;
 import java.util.UUID;
