+++ src/main/java/com/weeklyMission/console/ConsoleIO.java
@@ -1,6 +1,6 @@
 package com.weeklyMission.console;
 
-import com.weeklyMission.Voucher.dto.VoucherResponse;
+import com.weeklyMission.voucher.dto.VoucherResponse;
 import java.util.List;
 import java.util.Scanner;
 import org.springframework.stereotype.Component;
