+++ src/main/java/com/weeklyMission/voucher/domain/PercentDiscountVoucher.java
@@ -1,13 +1,13 @@
-package com.weeklyMission.Voucher.domain;
+package com.weeklyMission.voucher.domain;
 
 import java.util.UUID;
 
-public class FixedAmountVoucher implements Voucher{
+public class PercentDiscountVoucher implements Voucher{
 
     private final UUID voucherId;
     private final long amount;
 
-    public FixedAmountVoucher(UUID voucherId, long amount) {
+    public PercentDiscountVoucher(UUID voucherId, long amount) {
         this.voucherId = voucherId;
         this.amount = amount;
     }
@@ -24,6 +24,6 @@ public class FixedAmountVoucher implements Voucher{
 
     @Override
     public long discount(long beforeDiscount) {
-        return beforeDiscount-amount;
+        return beforeDiscount/amount;
     }
 }
